+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/pusher/DctiKafkaPusher.java
@@ -4,7 +4,6 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.annotation.JSONField;
 import com.dwarfeng.dcti.sdk.util.DataInfoUtil;
 import com.dwarfeng.dcti.stack.bean.dto.DataInfo;
-import com.dwarfeng.fdr.impl.handler.Pusher;
 import com.dwarfeng.fdr.stack.bean.entity.FilteredValue;
 import com.dwarfeng.fdr.stack.bean.entity.PersistenceValue;
 import com.dwarfeng.fdr.stack.bean.entity.RealtimeValue;
@@ -28,7 +27,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 
 /**
  * 标准数据接口Kafka推送器。
@@ -37,9 +35,9 @@ import java.util.Objects;
  * @since 1.5.0
  */
 @Component
-public class DctiKafkaPusher implements Pusher {
+public class DctiKafkaPusher extends AbstractPusher {
 
-    public static final String SUPPORT_TYPE = "dcti.kafka";
+    public static final String PUSHER_TYPE = "dcti.kafka";
 
     @Autowired
     @Qualifier("dctiKafkaPusher.kafkaTemplate")
@@ -54,9 +52,8 @@ public class DctiKafkaPusher implements Pusher {
     @Value("${pusher.dcti.kafka.topic.persistence_recorded}")
     private String persistenceRecordedTopic;
 
-    @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
+    public DctiKafkaPusher() {
+        super(PUSHER_TYPE);
     }
 
     @Override
@@ -117,6 +114,18 @@ public class DctiKafkaPusher implements Pusher {
     @Transactional(transactionManager = "dctiKafkaPusher.kafkaTransactionManager")
     public void persistenceRecorded(List<PersistenceValue> persistenceValues) {
         persistenceValues.forEach(this::persistenceRecorded);
+    }
+
+    @Override
+    public String toString() {
+        return "DctiKafkaPusher{" +
+                "kafkaTemplate=" + kafkaTemplate +
+                ", dataFilteredTopic='" + dataFilteredTopic + '\'' +
+                ", dataTriggeredTopic='" + dataTriggeredTopic + '\'' +
+                ", realtimeUpdatedTopic='" + realtimeUpdatedTopic + '\'' +
+                ", persistenceRecordedTopic='" + persistenceRecordedTopic + '\'' +
+                ", pusherType='" + pusherType + '\'' +
+                '}';
     }
 
     @Configuration