+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/pusher/MultiPusher.java
@@ -15,7 +15,6 @@ import org.springframework.stereotype.Component;
 import javax.annotation.PostConstruct;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.StringTokenizer;
 
 /**
@@ -25,9 +24,9 @@ import java.util.StringTokenizer;
  * @since 1.4.0
  */
 @Component
-public class MultiPusher implements Pusher {
+public class MultiPusher extends AbstractPusher {
 
-    public static final String SUPPORT_TYPE = "multi";
+    public static final String PUSHER_TYPE = "multi";
     private static final Logger LOGGER = LoggerFactory.getLogger(MultiPusher.class);
 
     @Autowired
@@ -38,6 +37,10 @@ public class MultiPusher implements Pusher {
 
     private final List<Pusher> delegates = new ArrayList<>();
 
+    public MultiPusher() {
+        super(PUSHER_TYPE);
+    }
+
     @PostConstruct
     public void init() throws HandlerException {
         StringTokenizer st = new StringTokenizer(delegateTypes, ",");
@@ -46,11 +49,6 @@ public class MultiPusher implements Pusher {
             delegates.add(pushers.stream().filter(p -> p.supportType(delegateType)).findAny()
                     .orElseThrow(() -> new HandlerException("未知的 pusher 类型: " + delegateType)));
         }
-    }
-
-    @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
     }
 
     @Override
@@ -139,5 +137,13 @@ public class MultiPusher implements Pusher {
                 LOGGER.warn("代理推送器推送数据失败，异常信息如下: ", e);
             }
         }
+    }
+
+    @Override
+    public String toString() {
+        return "MultiPusher{" +
+                "delegateTypes='" + delegateTypes + '\'' +
+                ", pusherType='" + pusherType + '\'' +
+                '}';
     }
 }