+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/pusher/PartialDrainPusher.java
@@ -12,16 +12,15 @@ import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
 import java.util.List;
-import java.util.Objects;
 
 /**
  * @author DwArFeng
  * @since 1.4.0
  */
 @Component
-public class PartialDrainPusher implements Pusher {
+public class PartialDrainPusher extends AbstractPusher {
 
-    public static final String SUPPORT_TYPE = "partial_drain";
+    public static final String PUSHER_TYPE = "partial_drain";
 
     @Autowired
     private List<Pusher> pushers;
@@ -39,15 +38,14 @@ public class PartialDrainPusher implements Pusher {
 
     private Pusher delegate;
 
+    public PartialDrainPusher() {
+        super(PUSHER_TYPE);
+    }
+
     @PostConstruct
     public void init() throws HandlerException {
         this.delegate = pushers.stream().filter(p -> p.supportType(delegateType)).findAny()
                 .orElseThrow(() -> new HandlerException("未知的 pusher 类型: " + delegateType));
-    }
-
-    @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
     }
 
     @Override
@@ -145,4 +143,18 @@ public class PartialDrainPusher implements Pusher {
     public void setDrainTriggeredValue(boolean drainTriggeredValue) {
         this.drainTriggeredValue = drainTriggeredValue;
     }
+
+    @Override
+    public String toString() {
+        return "PartialDrainPusher{" +
+                "pushers=" + pushers +
+                ", delegateType='" + delegateType + '\'' +
+                ", drainRealtimeValue=" + drainRealtimeValue +
+                ", drainPersistenceValue=" + drainPersistenceValue +
+                ", drainFilteredValue=" + drainFilteredValue +
+                ", drainTriggeredValue=" + drainTriggeredValue +
+                ", pusherType='" + pusherType + '\'' +
+                '}';
+    }
+
 }